LayoutTests/inspector/debugger/break-on-exception-throw-in-promise-expected.txt

CONSOLE MESSAGE: line 36: testThrowingInPromise
CONSOLE MESSAGE: line 38: in promise
CONSOLE MESSAGE: line 7: throwing TestError
Checking pause locations when pausing on all exceptions.

PAUSE #1 AT: doThrow:8:22
